fix: :bug: incorrect component order on mobile
