+++ apps/www/app/docs/components/page.tsx
@@ -1,13 +1,9 @@
 import React from "react"
 import type { Metadata } from "next"
-import Link from "next/link"
-import { Index } from "@/__registry__"
 import Balancer from "react-wrap-balancer"
 
-import { docsConfig } from "@/config/docs"
 import { siteConfig } from "@/config/site"
 import { absoluteUrl } from "@/lib/utils"
-import { ComponentPreview } from "@/components/component-preview"
 import {
   Breadcrumb,
   BreadcrumbItem,
@@ -15,16 +11,8 @@ import {
   BreadcrumbList,
   BreadcrumbSeparator,
 } from "@/registry/default/ui/breadcrumb"
-import {
-  Card,
-  CardContent,
-  CardDescription,
-  CardHeader,
-  CardTitle,
-} from "@/registry/default/ui/card"
-import { registry } from "@/registry/registry"
 
-import PreviewComponentOverview from "./previews"
+import GridComponents from "./grid-components"
 
 export async function generateMetadata(): Promise<Metadata> {
   return {
@@ -83,57 +71,8 @@ export default function ComponentsPage() {
             </Balancer>
           </p>
         </div>
-        {/* <div className="grid grid-cols-1 gap-3 pb-12 pt-8 md:grid-cols-2 md:gap-4"> */}
-        <div className="mt-4 grid grid-cols-1 gap-3 xl:grid-cols-2 xl:gap-4 2xl:gap-6">
-          <GridListPreview />
-          <GridListPreview side="right" />
-        </div>
+        <GridComponents />
       </div>
     </main>
-  )
-}
-
-function GridListPreview({ side = "left" }: { side?: "left" | "right" }) {
-  const getCompLabel = (componentName: string) => {
-    return Object.values(docsConfig.sidebarNav[1].items).find(
-      (item) => item.title.toLowerCase().split(" ").join("-") === componentName
-    )?.label
-  }
-
-  return (
-    <div className="flex flex-col gap-3 xl:gap-4 2xl:gap-6">
-      {registry
-        .filter((component) => component.type === "components:ui")
-        .filter((_, index) => (index + (side === "right" ? 1 : 0)) % 2 === 0)
-        .map((component, index) => (
-          <Card
-            key={side + index + component.name}
-            className="relative flex h-min flex-col transition-colors hover:bg-accent/70"
-          >
-            <Link
-              className="absolute inset-0 text-transparent"
-              href={`/docs/components/${component.name}`}
-            >
-              Link
-            </Link>
-            <CardHeader className="pointer-events-none">
-              <CardTitle className="flex items-center gap-2">
-                <span>
-                  {component.name.charAt(0).toLocaleUpperCase() +
-                    component.name.slice(1)}
-                </span>
-                {getCompLabel(component.name) && (
-                  <span className="ml-2 rounded-md bg-[#adfa1d] px-1.5 py-0.5 text-xs leading-none text-[#000000] no-underline group-hover:no-underline">
-                    {getCompLabel(component.name)}
-                  </span>
-                )}
-              </CardTitle>
-            </CardHeader>
-            <CardContent className="pointer-events-none flex min-h-24 flex-1 items-center justify-center">
-              <PreviewComponentOverview name={`${component.name}-demo`} />
-            </CardContent>
-          </Card>
-        ))}
-    </div>
   )
 }
